pyproject.toml
[project]
name = "modeling_player-specific_chess_behaviors"
version = "1.0.0"
description = ""
readme = "README.md"
requires-python = "==3.13.0"
dependencies = [
    "chess>=1.11.2",
    "cucim-cu12==26.2.*",
    "cudf-cu12==26.2.*",
    "cugraph-cu12==26.2.*",
    "cuml-cu12==26.2.*",
    "cuvs-cu12==26.2.*",
    "cuxfilter-cu12==26.2.*",
    "dask-cudf-cu12==26.2.*",
    "einops>=0.8.2",
    "fastparquet>=2026.3.0",
    "flask>=3.1.3",
    "gdown>=5.2.1",
    "lxml>=6.0.2",
    "maia2>=0.9",
    "nx-cugraph-cu12==26.2.*",
    "polars>=1.39.3",
    "pyarrow>=23.0.1",
    "pydantic>=2.12.5",
    "pylibraft-cu12==26.2.*",
    "pyyaml>=6.0.3",
    "pyzstd>=0.19.1",
    "raft-dask-cu12==26.2.*",
    "requests>=2.28.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "tenacity>=9.1.4",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "umap-learn>=0.5.11",
]

[[tool.uv.index]]
url = "https://pypi.nvidia.com/"

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu126"
